git: name-rev
